Use regexp flags via helper
