+++ packages/core-js/internals/check-correctness-of-regexp-flags.js
@@ -1,0 +1,49 @@
+'use strict';
+var globalThis = require('./global-this');
+var DESCRIPTORS = require('./descriptors');
+var fails = require('./fails');
+
+// babel-minify and Closure Compiler transpiles RegExp('.', 'd') -> /./d and it causes SyntaxError
+var RegExp = globalThis.RegExp;
+var RegExpPrototype = RegExp.prototype;
+
+module.exports = function () {
+  return DESCRIPTORS && !fails(function () {
+    var INDICES_SUPPORT = true;
+    try {
+      RegExp('.', 'd');
+    } catch (error) {
+      INDICES_SUPPORT = false;
+    }
+
+    var O = {};
+    // modern V8 bug
+    var calls = '';
+    var expected = INDICES_SUPPORT ? 'dgimsy' : 'gimsy';
+
+    var addGetter = function (key, chr) {
+      // eslint-disable-next-line es/no-object-defineproperty -- safe
+      Object.defineProperty(O, key, { get: function () {
+        calls += chr;
+        return true;
+      } });
+    };
+
+    var pairs = {
+      dotAll: 's',
+      global: 'g',
+      ignoreCase: 'i',
+      multiline: 'm',
+      sticky: 'y'
+    };
+
+    if (INDICES_SUPPORT) pairs.hasIndices = 'd';
+
+    for (var key in pairs) addGetter(key, pairs[key]);
+
+    // eslint-disable-next-line es/no-object-getownpropertydescriptor -- safe
+    var result = Object.getOwnPropertyDescriptor(RegExpPrototype, 'flags').get.call(O);
+
+    return result !== expected || calls !== expected;
+  });
+};
