+++ packages/core-js/internals/detection-of-correctness-regexp-flags.js
@@ -1,0 +1,3 @@
+'use strict';
+
+module.exports = { correct: false };
