+++ packages/core-js/modules/es.string.match-all.js
@@ -12,7 +12,7 @@ var anObject = require('../internals/an-object');
 var isObject = require('../internals/is-object');
 var classof = require('../internals/classof-raw');
 var isRegExp = require('../internals/is-regexp');
-var getRegExpFlags = require('../internals/regexp-get-flags');
+var getRegExpFlags = require('../internals/get-regexp-flags');
 var getMethod = require('../internals/get-method');
 var defineBuiltIn = require('../internals/define-built-in');
 var fails = require('../internals/fails');
