+++ packages/core-js/modules/es.string.replace-all.js
@@ -8,7 +8,7 @@ var isObject = require('../internals/is-object');
 var isRegExp = require('../internals/is-regexp');
 var toString = require('../internals/to-string');
 var getMethod = require('../internals/get-method');
-var getRegExpFlags = require('../internals/regexp-get-flags');
+var getRegExpFlags = require('../internals/get-regexp-flags');
 var getSubstitution = require('../internals/get-substitution');
 var wellKnownSymbol = require('../internals/well-known-symbol');
 var IS_PURE = require('../internals/is-pure');
