+++ packages/core-js/modules/web.structured-clone.js
@@ -19,7 +19,7 @@ var createProperty = require('../internals/create-property');
 var createNonEnumerableProperty = require('../internals/create-non-enumerable-property');
 var lengthOfArrayLike = require('../internals/length-of-array-like');
 var validateArgumentsLength = require('../internals/validate-arguments-length');
-var getRegExpFlags = require('../internals/regexp-get-flags');
+var getRegExpFlags = require('../internals/get-regexp-flags');
 var MapHelpers = require('../internals/map-helpers');
 var SetHelpers = require('../internals/set-helpers');
 var setIterate = require('../internals/set-iterate');
